refactoring : 컴포넌트 분리 Add : 로그인 인증, mockdata
