+++ public/data/commentData.json
@@ -1,32 +1,37 @@
 [
-    {
-      "id": 1,
-      "userName": "wecode",
-      "content": "Welcome to world best coding bootcamp!",
-      "isLiked": true
-    },
-    {
-      "id": 2,
-      "userName": "joonsikyang",
-      "content": "Hi there.",
-      "isLiked": false
-    },
-    {
-      "id": 3,
-      "userName": "jayPark",
-      "content": "Hey.",
-      "isLiked": false
-    },
-    {
-        "id": 4,
-        "userName": "sanghyeon",
-        "content": "Hey.hey hey",
-        "isLiked": false
-    },
-    {
-        "id": 5,
-        "userName": "hanho",
-        "content": "what!?",
-        "isLiked": false
-    }
+  {
+    "id": 101,
+    "userName": "wecode",
+    "content": "Welcome to world best coding bootcamp!",
+    "isLiked": true,
+    "feedID": 1
+  },
+  {
+    "id": 102,
+    "userName": "joonsikyang",
+    "content": "Hi there.",
+    "isLiked": false,
+    "feedID": 2
+  },
+  {
+    "id": 103,
+    "userName": "jayPark",
+    "content": "Hey.",
+    "isLiked": false,
+    "feedID": 3
+  },
+  {
+    "id": 104,
+    "userName": "sanghyeon",
+    "content": "Hey.hey hey",
+    "isLiked": false,
+    "feedID": 4
+  },
+  {
+    "id": 105,
+    "userName": "hanho",
+    "content": "what!?",
+    "isLiked": false,
+    "feedID": 5
+  }
 ]
