+++ public/data/feedData.json
@@ -7,7 +7,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -18,7 +18,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -29,7 +29,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -40,7 +40,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -51,7 +51,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -62,7 +62,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -73,7 +73,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -84,7 +84,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -95,7 +95,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   },
   {
@@ -106,7 +106,7 @@
     "profileUrl": "",
     "feedUrl": "",
     "whoHeart": ["sanghyeon"],
-    "heartCnt": 0,
+    "heartCnt": 1,
     "isLiked": false
   }
 ]
